requirements.txt
numpy
mxnet-cu90
matplotlib
pillow
scipy
pandas
nltk